修改数据

+++ capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/pojo/SchoolGoldFactory.java
@@ -33,7 +33,7 @@ public class SchoolGoldFactory implements Serializable {
     /**
      * 金币数量
      */
-    private Integer gold;
+    private Double gold;
 
 
 }
